FRONTEND/CODE/requirements.txt
absl-py==2.2.1
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fastjsonschema==2.21.1
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.4.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
ipython-genutils==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mysql-connector-python==9.2.0
namex==0.0.8
narwhals==1.33.0
nbformat==4.2.0
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.3
pytz==2025.2
pywin32==308
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
sounddevice==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.14.1
xgboost==3.0.0
